CapstoneProject/my_nlp_app/requirements.txt
flask
requests
nltk
matplotlib
fasttext
keybert
flask_sqlalchemy
python-dotenv
pymysql
flask_cors